Moving method and adding torch type hints

+++ src/readability_classifier/models/encoders/dataset_utils.py
@@ -90,10 +90,14 @@ def load_encoded_dataset(data_dir: str) -> ReadabilityDataset:
 
     # Convert loaded data to torch.Tensors
     for sample in dataset_list:
-        sample["matrix"] = torch.tensor(sample["matrix"], dtype=torch.float32)
-        sample["input_ids"] = torch.tensor(sample["input_ids"], dtype=torch.long)
+        sample["matrix"] = torch.tensor(
+            sample["matrix"], dtype=torch.float32
+        )  # Why not int?
+        sample["input_ids"] = torch.tensor(
+            sample["input_ids"], dtype=torch.float32
+        )  # Why not int?
         sample["token_type_ids"] = torch.tensor(
-            sample["token_type_ids"], dtype=torch.long
+            sample["token_type_ids"], dtype=torch.float32  # Why not int?
         )
         sample["image"] = torch.tensor(sample["image"], dtype=torch.float32)
         sample["score"] = torch.tensor(sample["score"], dtype=torch.float32)
